cleanup: simplify prop definitions
